Addressing comments - moving retryPolicy to spi
