+++ pinot-batch-ingestion/pinot-standalone/src/main/java/org/apache/pinot/ingestion/standalone/SegmentUriPushJobRunner.java
@@ -29,10 +29,10 @@ import org.apache.commons.configuration.MapConfiguration;
 import org.apache.pinot.common.exception.HttpErrorStatusException;
 import org.apache.pinot.common.utils.FileUploadDownloadClient;
 import org.apache.pinot.common.utils.SimpleHttpResponse;
-import org.apache.pinot.common.utils.retry.AttemptsExceededException;
-import org.apache.pinot.common.utils.retry.RetriableOperationException;
-import org.apache.pinot.common.utils.retry.RetryPolicies;
-import org.apache.pinot.filesystem.PinotFSFactory;
+import org.apache.pinot.spi.utils.retry.AttemptsExceededException;
+import org.apache.pinot.spi.utils.retry.RetriableOperationException;
+import org.apache.pinot.spi.utils.retry.RetryPolicies;
+import org.apache.pinot.spi.filesystem.PinotFSFactory;
 import org.apache.pinot.ingestion.common.Constants;
 import org.apache.pinot.ingestion.common.PinotClusterSpec;
 import org.apache.pinot.ingestion.common.PinotFSSpec;
