+++ pinot-controller/src/main/java/org/apache/pinot/controller/ControllerConf.java
@@ -27,7 +27,7 @@ import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.PropertiesConfiguration;
 import org.apache.pinot.common.protocols.SegmentCompletionProtocol;
 import org.apache.pinot.common.utils.StringUtil;
-import org.apache.pinot.filesystem.LocalPinotFS;
+import org.apache.pinot.spi.filesystem.LocalPinotFS;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
