+++ pinot-spi/src/main/java/org/apache/pinot/spi/utils/retry/RetryPolicy.java
@@ -16,14 +16,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.utils.retry;
+package org.apache.pinot.spi.utils.retry;
+
+import java.util.concurrent.Callable;
+
 
 /**
- * The <code>RetriableOperationException</code> indicates that the retriable operation threw an exception.
+ * Retry policy, encapsulating the logic needed to retry an operation until it succeeds.
  */
-public class RetriableOperationException extends AttemptFailureException {
+public interface RetryPolicy {
 
-  public RetriableOperationException(Throwable cause) {
-    super(cause);
-  }
+  /**
+   * Attempts to do the operation until it succeeds, aborting if an exception is thrown by the operation or number of
+   * attempts exhausted.
+   *
+   * @param operation The operation to attempt, which returns true on success and false on failure.
+   * @throws AttemptsExceededException
+   * @throws RetriableOperationException
+   */
+  void attempt(Callable<Boolean> operation)
+      throws AttemptsExceededException, RetriableOperationException;
 }
