+++ pinot-tools/src/main/java/org/apache/pinot/tools/PinotIdealstateChanger.java
@@ -23,7 +23,7 @@ import javax.annotation.Nullable;
 import org.apache.helix.HelixManager;
 import org.apache.helix.model.IdealState;
 import org.apache.pinot.common.utils.helix.HelixHelper;
-import org.apache.pinot.common.utils.retry.RetryPolicies;
+import org.apache.pinot.spi.utils.retry.RetryPolicies;
 
 
 public class PinotIdealstateChanger extends PinotZKChanger {
